refactor: remove interfaces from hooks
